pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "voicedebate"
version = "0.1.0"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
description = "An AI-powered voice debate application"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "kivymd>=1.1.1",
    "kivy>=2.2.1",
    "deepgram-sdk>=2.11.0",
    "elevenlabs>=0.2.24",
    "anthropic>=0.5.0",
    "openai>=1.3.0",
    "asyncpg>=0.28.0",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.2",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/voice-debate"
"Bug Tracker" = "https://github.com/yourusername/voice-debate/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["voicedebate*"]
namespaces = false
